+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Don't read environment variables without prefix
 - Keep consistency of local storage when removing a backup
 - Tool version automatically defined in build scripts
+- TAR header path was including volume name in Windows environments
 
 ### Changed
 - Tool moved to cmd subdirectory